fix(tests): resolve 89 verified inspection defects across the test suite
Second pass of the ReSharper triage, over native/tests. 89 fixes applied, 145
findings confirmed stale against the current tree, and 350 declined.

The declines are the point of this commit as much as the fixes. A large share of
what the analyser suggested here would have silently deleted coverage:

* performance-unnecessary-copy-initialization wanted deliberate deep copies
  turned into references in game_data_service_tests, frontend_timeline_tests and
  the TDX candidate-image tests. Those copies are taken before the backing
  DATA.HOG is overwritten on disk, precisely to prove the IR owns its bytes;
  aliasing them would make the paired assertion compare an object with itself
  and quietly destroy three ownership proofs.
* readability-container-size-empty wanted `empty == Buffer{}` rewritten as
  `empty.empty()`, which drops all operator== coverage.
* The bugprone-use-after-move hits in input_trace_tests and
  frontend_presentation_tests ARE the tests -- they assert the documented inert
  state of a moved-from object.
* clang-diagnostic-float-equal in frontend_texture_sampler_tests is deliberate:
  the two lookups must be bit-identical, and NearlyEqual would weaken it.
* CppUseInternalLinkage on three *FailureCount functions in the POP tests would
  have broken the link outright -- hog_archive_tests declares and calls them.

What was fixed was real. The most serious was undefined behaviour in the stdout
capture helper every stream assertion in frontend_envelope_commands_tests runs
through: `original_(stream.rdbuf(buffer_.rdbuf()))` called buffer_.rdbuf() on an
ostringstream that had not been constructed yet, because buffer_ was declared
after original_. Also fixed: implementation-defined size_t to difference_type
narrowing in hog_archive_tests iterator arithmetic; several 32-bit
multiplications widened only after overflow could occur; two test functions that
had external linkage solely because a namespace-closing brace sat in the wrong
place; and a handful of verified-unused includes.

Fixture bytes were treated as data throughout -- no synthetic or malformed-input
fixture was reformatted or altered. Independent adversarial review of every
diff, looking specifically for removed assertions, changed expected values and
altered fixtures, found no blocking or important problem. Build clean under
/W4 /WX, ctest 121/121, both project gates pass.

+++ native/tests/frontend_tdx_decoder_tests.cpp
@@ -166,7 +166,7 @@ void WriteTransferControlPrefix(std::vector<std::byte>& bytes, const std::size_t
     WritePacket(bytes, kPaletteObject, base_pointer, 1U, palette_width, palette_height,
                 palette_bytes);
 
-    for (std::uint32_t index = 0; index < palette_bytes / 4U; ++index)
+    for (std::size_t index = 0; index < palette_bytes / 4U; ++index)
     {
         bytes[kPaletteData + index * 4U] = static_cast<std::byte>(index & 0xFFU);
         bytes[kPaletteData + index * 4U + 1U] = std::byte{0x22};
